pyproject.toml

[project]
name = "sqse-aead"
version = "0.1.0"
description = "SQSE AEAD: GPU/DLL-gestützte Sub-Quantum State Encryption mit Argon2id, Nonce/Counter, BLAKE3-MAC"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Ralf Krümmel" }]
dependencies = [
  "numpy>=2.0.0",
  "blake3>=0.3.3",
  "typer>=0.12.3",
  "rich>=13.7.1",
]

[project.optional-dependencies]
password = ["argon2-cffi>=23.1.0"]
app = ["streamlit>=1.36.0"]

[project.scripts]
sqse-aead = "sqse_aead_pkg.cli:app"
